code/libs/python/momo-mom/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "momo_mom" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../../../coverage/code/libs/python/momo-mom/html'
 --cov-report xml:'../../../../coverage/code/libs/python/momo-mom/coverage.xml'
 --html='../../../../reports/code/libs/python/momo-mom/unittests/html/index.html'
 --junitxml='../../../../reports/code/libs/python/momo-mom/unittests/junit.xml'
"""

[project]
name = "momo-mom"
version = "1.0.0"
description = "Universal command mapping system for AI-friendly developer tools"
requires-python = ">=3.9,<4"
readme = 'README.md'
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0.0",
    "pytest-cov>=6.2.1",
]

[project.scripts]
mom = "momo_mom.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["momo_mom"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
  "autopep8>=2.3.1",
  "ruff>=0.8.2",
  "pytest>=8.3.4",
  "pytest-sugar>=1.0.0",
  "pytest-cov>=6.0.0",
  "pytest-html>=4.1.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = [
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "dist",
]

line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]
ignore = []

fixable = ["ALL"]
unfixable = []
